Refactor table functions into Table module and then delegate functions
