+++ lib/harmonium/pagination.ex
@@ -3,63 +3,65 @@ defmodule Harmonium.Pagination do
   @page_links_to_show 2
   import Phoenix.HTML
 
+  def pagination(_conn, _url, _page_number, 1) do
+    nil
+  end
+
   def pagination(conn, url, page_number, total_pages) do
     ~E"""
-    <%= if show_pagination?(total_pages) do %>
     <div class="rev-PaginationWrapper text-center">
-    <ul class="rev-Pagination" role="navigation" aria-label="Pagination">
-      <%= unless first_page?(page_number) do %>
-        <li class="rev-Pagination-arrow">
-          <a href="<%= make_page_link(conn, 1, url) %>">
-            <span><i class="icon-angle-double-left"></i>First<span class="ShowForSR"> page</span></span>
-          </a>
-        </li>
-        <li class="rev-Pagination-arrow">
-          <a href="<%= make_previous_page_link(conn, page_number, url) %>">
-            <span><i class="icon-angle-left"></i>Previous<span class="ShowForSR"> page</span></span>
-          </a>
-        </li>
-      <% end %>
+      <ul class="rev-Pagination" role="navigation" aria-label="Pagination">
+        <%= unless first_page?(page_number) do %>
+          <li class="rev-Pagination-arrow">
+            <a href="<%= make_page_link(conn, 1, url) %>">
+              <span><i class="icon-angle-double-left"></i>First<span class="ShowForSR"> page</span></span>
+            </a>
+          </li>
+          <li class="rev-Pagination-arrow">
+            <a href="<%= make_previous_page_link(conn, page_number, url) %>">
+              <span><i class="icon-angle-left"></i>Previous<span class="ShowForSR"> page</span></span>
+            </a>
+          </li>
+        <% end %>
 
-      <%= for page <- 1..total_pages do %>
-        <%= if show_previous_ellipsis?(page, page_number) do %>
-          <li class="rev-Pagination-dots">
-            ...
+        <%= for page <- 1..total_pages do %>
+          <%= if show_previous_ellipsis?(page, page_number) do %>
+            <li class="rev-Pagination-dots">
+              ...
+            </li>
+          <% end %>
+          <%= if show_page_link?(page, page_number) do %>
+            <li class="rev-Pagination-number <%= selected_page_class(page, page_number)%>">
+              <%= if page == page_number do %>
+                <span class="ShowForSR">You're on page </span><a><%= page_number %></a>
+              <% else %>
+                <a href="<%= make_page_link(conn, page, url) %>" aria-label="Page <%= page %>"><%= page %></a>
+              <% end %>
+            </li>
+          <% end %>
+          <%= if show_next_ellipsis?(page, page_number, total_pages) do %>
+            <li class="rev-Pagination-dots">
+              ...
+            </li>
+          <% end %>
+        <% end %>
+        <%= unless last_page?(page_number, total_pages) do %>
+          <li class="rev-Pagination-arrow">
+            <a href="<%= make_next_page_link(conn, page_number, url) %>">
+              <span>Next<span class="ShowForSR"> page</span><i class="icon-angle-right"></i></span>
+            </a>
+          </li>
+          <li class="rev-Pagination-arrow">
+            <a href="<%= make_page_link(conn, total_pages, url) %>">
+              <span>Last<span class="ShowForSR"> page</span><i class="icon-angle-double-right"></i></span>
+            </a>
           </li>
         <% end %>
-        <%= if show_page_link?(page, page_number) do %>
-          <li class="rev-Pagination-number <%= selected_page_class(page, page_number)%>">
-            <%= if page == page_number do %>
-              <span class="ShowForSR">You're on page </span><a><%= page_number %></a>
-            <% else %>
-              <a href="<%= make_page_link(conn, page, url) %>" aria-label="Page <%= page %>"><%= page %></a>
-            <% end %>
-          </li>
-        <% end %>
-        <%= if show_next_ellipsis?(page, page_number, total_pages) do %>
-          <li class="rev-Pagination-dots">
-            ...
-          </li>
-        <% end %>
-      <% end %>
-      <%= unless last_page?(page_number, total_pages) do %>
-        <li class="rev-Pagination-arrow">
-          <a href="<%= make_next_page_link(conn, page_number, url) %>">
-            <span>Next<span class="ShowForSR"> page</span><i class="icon-angle-right"></i></span>
-          </a>
-        </li>
-        <li class="rev-Pagination-arrow">
-          <a href="<%= make_page_link(conn, total_pages, url) %>">
-            <span>Last<span class="ShowForSR"> page</span><i class="icon-angle-double-right"></i></span>
-          </a>
-        </li>
-      <% end %>
-    </ul>
-    <div class="rev-PaginationWrapper-pageList">
-      <span class="Small">( Page <%= page_number %> of <%= total_pages %> )</span>
+      </ul>
+      <div class="rev-PaginationWrapper-pageList">
+        <span class="Small">( Page <%= page_number %> of <%= total_pages %> )</span>
+      </div>
     </div>
-    </div>
-    <% end %>
     """
   end
 
@@ -95,9 +97,6 @@ defmodule Harmonium.Pagination do
 
   defp last_page?(page_number, total_pages) when page_number == total_pages, do: true
   defp last_page?(_, _), do: false
-
-  defp show_pagination?(1), do: false
-  defp show_pagination?(_), do: true
 
   defp selected_page_class(page_number, current_page) when page_number == current_page do
     "rev-Pagination-number--selected"
